@@ -1,6 +1,6 @@
 {
   "name": "mozaik-ext-github",
-  "version": "1.0.6",
+  "version": "2.0.0",
   "description": "Mozaik github widgets",
   "repository": {
     "type": "git",
@@ -14,41 +14,44 @@
   "homepage": "https://github.com/plouc/mozaik-ext-github",
   "main": "./src/components/index.js",
   "engines": {
-    "node": "0.12.x"
+    "node": ">=5.0.0",
+    "npm" : ">=3.0.0"
   },
   "dependencies": {
-    "babelify": "^6.1.3",
     "chalk": "^1.1.0",
     "convict": "^0.6.1",
-    "lodash": "^3.2.0",
+    "lodash": "4.6.1",
     "moment": "^2.10.3",
     "moment-duration-format": "^1.3.0",
-    "react": "^0.13.3",
+    "react": "0.14.7",
     "react-mixin": "^1.7.0",
-    "reflux": "^0.2.5",
+    "reflux": "0.4.0",
     "superagent": "^1.2.0",
     "superagent-bluebird-promise": "^0.6.0"
   },
   "devDependencies": {
-    "babel": "^5.8.12",
-    "babel-eslint": "^1.0.6",
-    "chai": "^3.2.0",
-    "eslint": "^0.15.1",
+    "babel-core": "6.7.2",
+    "babel-eslint": "5.0.0",
+    "babel-preset-es2015": "6.6.0",
+    "babel-preset-react": "6.5.0",
+    "babel-preset-stage-0": "6.5.0",
+    "babel-register": "6.7.2",
+    "enzyme": "2.1.0",
+    "eslint": "2.2.0",
+    "eslint-plugin-react": "4.2.2",
+    "expect": "1.15.2",
     "jsdom": "^2.0.0",
-    "mocha": "^2.2.5",
-    "mockery": "^1.4.0",
-    "sinon": "^1.15.4"
+    "mocha": "2.4.5",
+    "mockery": "1.4.1",
+    "react-addons-test-utils": "0.14.7",
+    "react-dom": "0.14.7",
+    "sinon": "1.17.3"
   },
   "peerDependencies": {
-    "mozaik": "1.x"
+    "mozaik": "1.2.x"
   },
   "scripts": {
-    "test": "mocha --opts mocha.opts ./src/tests/**/*.test.js",
-    "eslint": "eslint ./src/*"
-  },
-  "browserify": {
-    "transform": [
-      "babelify"
-    ]
+    "eslint": "eslint --ext .js --ext .jsx ./src/*",
+    "test": "mocha --opts mocha.opts"
   }
 }
