@@ -1,1 +1,3 @@
-module.exports = require('./src/client');
+import client from './src/client';
+
+export default client;
