@@ -1,17 +1,22 @@
 import React, { Component, PropTypes } from 'react';
 import reactMixin                      from 'react-mixin';
 import { ListenerMixin }               from 'reflux';
-import _                               from 'lodash';
 import Branch                          from './Branch.jsx';
 import Mozaik                          from 'mozaik/browser';
 
 
 class Branches extends Component {
+    static displayName = 'Branches';
+
+    static propTypes = {
+        repository: PropTypes.string.isRequired,
+        title:      PropTypes.string
+    };
+
     constructor(props) {
         super(props);
-        this.state = {
-            branches: []
-        };
+
+        this.state = { branches: [] };
     }
 
     getApiRequest() {
@@ -19,16 +24,12 @@ class Branches extends Component {
 
         return {
             id:     `github.branches.${ repository }`,
-            params: {
-                repository: repository
-            }
+            params: { repository }
         };
     }
 
     onApiData(branches) {
-        this.setState({
-            branches: branches
-        });
+        this.setState({ branches });
     }
 
     render() {
@@ -37,7 +38,8 @@ class Branches extends Component {
 
         let titleNode = title === undefined ? (
             <span>
-                <span className="widget__header__subject">{repository}</span> branches
+                <span className="widget__header__subject">{repository}</span>&nbsp;
+                branches
             </span>
         ) : title;
 
@@ -51,21 +53,17 @@ class Branches extends Component {
                     <i className="fa fa-code-fork" />
                 </div>
                 <div className="widget__body">
-                    {branches.map(branch => {
-                        return <Branch key={branch.name} branch={branch}/>;
-                    })}
+                    {branches.map(branch => (
+                        <Branch key={branch.name} branch={branch} />
+                    ))}
                 </div>
             </div>
         );
     }
 }
 
-Branches.propTypes = {
-    repository: PropTypes.string.isRequired,
-    title:      PropTypes.string
-};
-
 reactMixin(Branches.prototype, ListenerMixin);
 reactMixin(Branches.prototype, Mozaik.Mixin.ApiConsumer);
 
-export { Branches as default };
+
+export default Branches;
