@@ -1,16 +1,20 @@
 import React, { Component, PropTypes } from 'react';
 import reactMixin                      from 'react-mixin';
 import { ListenerMixin }               from 'reflux';
-import _                               from 'lodash';
 import Mozaik                          from 'mozaik/browser';
 
 
 class UserBadge extends Component {
+    static displayName = 'UserBadge';
+
+    static propTypes = {
+        user: PropTypes.string.isRequired
+    };
+
     constructor(props) {
         super(props);
-        this.state = {
-            user: null
-        };
+
+        this.state = { user: null };
     }
 
     getApiRequest() {
@@ -18,51 +22,48 @@ class UserBadge extends Component {
 
         return {
             id:     `github.user.${ user }`,
-            params: {
-                user: user
-            }
+            params: { user }
         };
     }
 
     onApiData(user) {
-        this.setState({
-            user: user
-        });
+        this.setState({ user });
     }
 
     render() {
-        var userNode = (
+        let userNode = (
             <div className="widget__body" />
         );
 
-        if (this.state.user) {
+        const { user } = this.state;
+        if (user) {
             userNode = (
                 <div className="widget__body">
                     <div className="github__user-badge__banner">
                         <span className="github__user-badge__avatar">
-                            <img src={this.state.user.avatar_url} />
+                            <img src={user.avatar_url} />
                         </span>
                     </div>
                     <div className="github__user-badge__info">
                         <div className="github__user-badge__info__item">
-                            <span className="count">{this.state.user.public_repos}</span>&nbsp;
+                            <span className="count">{user.public_repos}</span>&nbsp;
                             public repos
                         </div>
                         <div className="github__user-badge__info__item">
-                            <span className="count">{this.state.user.public_gists}</span>&nbsp;
+                            <span className="count">{user.public_gists}</span>&nbsp;
                             public gists
                         </div>
                         <div className="github__user-badge__info__item">
-                            <span className="count">{this.state.user.followers}</span>&nbsp;
+                            <span className="count">{user.followers}</span>&nbsp;
                             followers
                         </div>
                         <div className="github__user-badge__info__item">
-                            <span className="count">{this.state.user.following}</span>&nbsp;
+                            <span className="count">{user.following}</span>&nbsp;
                             following
                         </div>
                         <div className="github__user-badge__info__item">
                             company:&nbsp;
-                            <span className="prop__value">{this.state.user.company}</span>
+                            <span className="prop__value">{user.company}</span>
                         </div>
                     </div>
                 </div>
@@ -81,11 +82,8 @@ class UserBadge extends Component {
     }
 }
 
-UserBadge.propTypes = {
-    user: PropTypes.string.isRequired
-};
-
 reactMixin(UserBadge.prototype, ListenerMixin);
 reactMixin(UserBadge.prototype, Mozaik.Mixin.ApiConsumer);
 
-export { UserBadge as default };
+
+export default UserBadge;
